requirements.txt
opencv-python~=4.11.0.86
numpy~=2.2.4
scikit-image~=0.25.2
tensorflow~=2.19.0
dahuffman~=0.4.2
pillow~=11.1.0
keras~=3.9.0
scikit-learn~=1.6.1
editdistance~=0.8.1